Add solution runner and puzzle input downloader
I've written a simple solution runner and puzzle input download for this
year's Advent of Code. It will automatically download the puzzle input
for the specified day (after verifying that we're past the announced
release date), create a default solution directory from a template, and
run the results.

There's also an option to time the two parts of each puzzle using the
`timeit` module.

For more information, run `python -m aoc --help`.