Remove explicit level on the logging handler
Setting an explicit level on the logging handler means that changing the
logging level on a logger alone may not be alone: The handler may still
filter out the lower logging levels.

As I have no need for such a complicated setup here, I'm removing the
explicit level from the handler, relying solely on the levels set in the
logging hierarchy.