Add solution for day 5 of the Advent of Code
With a bit of a delay before pushing my solution, here's day 5. It's not
as polished as I like, but there are two solutions that work.

A third solution that would probably be better than either of the two I
added is an arithmetic solution that determines the missing number based
on comparing the sum of IDs to what's missing from the sum of
consecutive natural numbers.