Fix Puzzle.time failing on prepare_function=None
Not every puzzle solution uses a `prepare_puzzle` function, which means
we should not try to run it if it doesn't exist. While the regular `run`
feature already took this into account, the `time` function tried to
time a non-existing preparation function anyway.

The reason was that while it checked for `function is None`, it lacked
the `continue` to make it skip that timing iteration.

I've added the missing `continue`.

In addition, I've also kaizened an unnecessary debug `print`.